API

Overview

HybridTreeBoosting.HTBinfo — Function

HTBinfo()

Basic information about the main functions in HTBoost (see help on each function for more details)

Setting up the model

HTBindexes_from_dates builds train and test sets indexes for expanding window cross-validation (if the user wants to over-ride the default block-cv)

HTBparam parameters, defaults or user-provided.

HTBdata y,x and, optionally, dates, weights, and names of features

Fitting and forecasting

HTBfit fits HTBoost with cv (or validation/early stopping) of number of trees and, optionally, other parameters; pre-determined options for more or less extensive cross-validation are controlled by :modality in HTBparam

HTBpredict predictions for y or natural parameter

HTBcv user-controlled cross-validation. (Note that the recommended process is to use :modality in HTBfit rather than HTBcv.)

POST-ESTIMATION ANALYSIS

HTBcoeff provides information on constant coefficients, e.g. dispersion and dof for loss=:t

HTBrelevance computes feature importance (Breiman et al 1984 relevance)

HTBoutput collects fitted parameters in matrices

HTBweightedtau computes weighted smoothing parameter to help assess function smoothness

HTBpartialplot partial dependence plots (keeping all other features fixed, not integrating out)

HTBmarginaleffect Numerical computation of marginal effects.

HTBplot_tau plotting sigmoid for a given tau

HTBplot_ppr provides tau and plot for projection pursuit for a single tree

Example of use of info function

help?> HTBinfo
or ...
julia> HTBinfo()

To find more information about a specific function, e.g. HTBfit

help?> HTBfit

Example of basic use of HTBoost functions with iid data and default settings

param  = HTBparam()                                 
data   = HTBdata(y,x,param)
output = HTBfit(data,param)
yf     = HTBpredict(x_test,output)

See the examples and tutorials for illustrations of use.

Setting up the model

HybridTreeBoosting.HTBparam — Type

HTBparam(;)

Parameters for HTBoost

Note: all Julia symbols can be replaced by strings. e.g. :L2 can be replaced by "L2".

Parameters that are most likely to be modified by user (all inputs are keywords with default values)

loss [:L2] Supported distributions:

:L2 (Gaussian), aliases :l2,:mse,:Gaussian,:normal

:logistic, aliases :binary (binary classification)

:multiclass (multiclass classification)

:t, aliases :student (student-t, robust alternative to :L2)

:Huber, aliases :huber

:gamma

:lognormal, aliases :logL2, :logl2 (positive continuous data)

:Poisson (count data)

:gammaPoisson, aliases :gammaPoisson,:gammapoisson,:negbin,:negative_binomial (aka negative binomial, count data)

:L2loglink, aliases :l2loglink (alternative to :L2 if y≥0)

:hurdleGamma (zero-inflated y)

:hurdleL2loglink (zero-inflated y)

:hurdleL2 (zero-inflated y)

See the examples for uses of each loss function. Fixed coefficients (such as shape for :gamma, dispersion and dof for :t, and overdispersion for :gammaPoisson) are computed internally by maximum likelihood. Inspect them using HTBcoeff(). In HTBpredict(), predictions are for E(y) if predict=:Ey (default), while predict=:Egamma forecasts the fitted parameter ( E(logit(prob) for :logistic, log(E(y)) for :gamma etc ... )

modality [:compromise] Options are: :accurate, :compromise (default), :fast, :fastest. These options are meant to replace the need for the user to cross-validate parameters. Advanced users with a big computational budget can still do so. :fast and :fastest run only one model, while :compromise and :accurate cross-validate the most important parameters. :fast runs only one model (only cv number of trees) at values defined in param = HTBparam(). :fastest runs only one model, setting lambda=0.2, nfold=1 and nofullsample=true (does not re-estimate on the full sample after cv). Recommended for faster preliminary analysis only. In most cases, :fast and :fastest also use the quadratic approximation to the loss for large samples. :compromise and :accurate cross-validates several models at the most important parameters (see HTBfit() for details), then stack all the cv models.

priortype [:hybrid] :hybrid encourages smoothness, but allows both smooth and sharp splits, :smooth forces smooth splits, :disperse is :hybrid but with no penalization encouraging smooth functions (not recommended). Set to :smooth if you want to force derivatives to be defined everywhere.

randomizecv [false] default is block-cv (aka purged cv); a time series or panel structure is automatically detected (see HTBdata) if a date column is provided. Set to true for standard cv.

nfold [4] n in n-fold cv. Set nfold = 1 for a single validation set (by default the last param.sharevalidation share of the sample). nfold, sharevalidation, and randomizecv are disregarded if train and test observations are provided by the user.

sharevalidation: [0.30] Can be: a) Integer, size of the validation set, or b) Float, share of validation set. Relevant only if nfold = 1.

indtrain_a:Vector{Vector{I}} [] for user's provided array of indices of train sets. e.g. vector of 5 vectors, each with indices of train set observations

indtest_a:Vector{Vector{I}} [] for user's provided array of indices of test sets. e.g. vector of 5 vectors, each with indices of train set observations

nofullsample [false] if true and nfold=1, HTBoost is not re-estimated on the full sample after validation. Reduces computing time by roughly 60%, at the cost of a modest loss of accuracy. Useful for very large datasets, in preliminary analysis, in simulations, and when instructions specify a train/validation split with no re-estimation on full sample. Activated by default when modality=:fastest.

cat_features [] vector of indices of categorical features, e.g. [2] or [2,5], or vector of names in DataFrame, e.g. [:wage,:age] or ["wage","age"]. If empty, categoricals are automatically detected as non-numerical features.

cv_categoricals [:default] whether to run preliminary cross-validation on parameters related to categorical features. :none uses default parameters :penalty runs a rough cv the penalty associated to the number of categories; recommended if n/n_cat if low for any feature, particularly if SNR is low :n0 runs a rough of cv the strength of the prior shrinking categorical values to the overall mean; recommended with highly unequal number of observations in different categories :both runs a rough cv or penalty and n0 :default uses :none for modality in [:fastest,:fast], :penalty for :compromise, and :both for :accurate

overlap: [0] number of overlaps in time series and panels. Typically overlap = h-1, where y(t) = Y(t+h)-Y(t). Used for purged-CV.

verbose [:Off] verbosity :On or :Off

warnings [:On] or :Off

Parameters that may sometimes be be modified by user

lambda [0.1 or 0.2] Learning rate. 0.1 for (nearly) best performance. 0.2 can be almost as accurate, particularly if the function is smooth and p is small. The default is 0.1, except in modality = :fastest, where it's 0.2. Modality = :compromise carries out the cv at lambda=0.2 and then fits the best model at 0.1. Consider 0.05 if tiny improvements in accuracy are important and computing time is not a concern.

depth [5] tree depth. Unless modality = :fast or :fastest, this is over-written as depth is cross-validated. See HTBfit() for more options.

weights NOTE: weights for weighted likelihood are set in HTBdata, not in HTBparam.

offset NOTE: offsets (aka exposures) are set in HTBdata, not in HTBparam. See examples/Offset or exposure.jl

sparsity_penalization [0.3] positive numbers encourage sparsity. The range [0.0-1.5] should cover most scenarios. Automatically cv in modality=:compromise and :accurate. Increase to obtain a more parsimonious model, set to 0 for standard boosting.

ntrees [3000] Maximum number of trees. HTBfit will automatically stop when cv loss stops decreasing.

sharevs [1.0] row subsampling in variable selection phase (only to choose feature on which to split.) Default is no subsampling. sharevs = :Auto sets the subsample size to min(n,50k*sqrt(n/50k)). At high n, sharevs<1 speeds up computations, but can reduce accuracy, particularly in sparse setting with low SNR.

subsampleshare_columns [1.0] column subsampling (aka feature subsampling) by level.

min_unique [:default] sharp splits are imposed on features with less than min_unique values (default is 5 for modality=:compromise or :accurate, else 10)

mixed_dc_sharp [false] true to force sharp splits on discrete and mixed discrete-continuous features (defined as having over 20% obs on a single value)

delete_missing [false] true to delete rows with missing values in any feature, false to handle missing internally (recommended).

theta [1] numbers larger than 1 imply tighter penalization on β (final leaf values) compared to default.

meanlntau [1.0] prior mean of log(τ). Set to higher numbers to suggest less smooth functions.

mugridpoints [11] number of points at which to evaluate μ during variable selection. 5 is sufficient on simulated data with normal or uniform distributions, but actual data may benefit from more (due to with highly non-Gaussian features). For extremely complex and nonlinear features, more than 10 may be needed.

force_sharp_splits [] optionally, a p vector of Bool, with j-th value set to true if the j-th feature is forced to enter with a sharp split.

force_smooth_splits [] optionally, a p vector of Bool, with j-th value set to true if the j-th feature is forced to enter with a smooth split (high values of τ not allowed).

cat_representation_dimension [4 (2 for classification)] 1 for mean encoding, 2 adds frequency, 3 adds variance, 4 adds robust skewness, 5 adds robust kurtosis

losscv [:default] loss function for cross-validation (:mse,:mae,:logistic,:sign).

n_refineOptim [10^6] MAXIMUM number of observations to use fit μ and τ (split point and smoothness). Lower numbers can provide speed-ups with very large n at some cost in terms of fit.

loglikdivide [1.0] Higher numbers increase the strength or all priors. The defaults sets it internally using HTBloglikdivide(), when it detects a dates series in HTBdata().

tau_threshold [10.0] lowest threshold for imposing sharp splits. Lower numbers give more sharp splits.

multiplier_stdtau [5.0] The default priors suggest smoother splits on features whose unconditional distribution (appropriately transformed according to the link function) is closer to the unconditional distribution of y or, when not applicable, to a Gaussian. To disengage this feature, set multiplier_stdtau = 0

Additional parameters to control the cross-validation process can be set in HTBfit(), but keeping the defaults is generally encouraged.

Example

param = HTBparam()

Example

param = HTBparam(nfold=1,nofullsample=true)

HybridTreeBoosting.HTBdata — Type

    HTBdata(y,x,param,[dates];weights=[],fnames=[],offset=[])

Collects and pre-processes data in preparation for fitting HTBoost

Inputs

y::Vector Vector of responses. Can be a vector of lables, or a dataframe with one column.

x Matrix of features. Can be a vector or matrix of floats, or a dataframe. Converted internally to a Matrix{T}, T as defined in HTBparam

param::HTBparam

Optional Inputs

dates::AbstractVector [1:n] Typically Vector{Date} or Vector{Int}. Used in cross-validation to determine splits. If not supplied, the default 1:n assumes a cross-section of independent realizations (conditional on x) or a single time series.

'weights' [ones(T,n)] vector of floats or Floats, weights for weighted likelhood

fnames::Vector{String} [x1, x2, ... ] feature names

'offset' vector of offsets (exposure), in logs if the loss adopts a loss link (:gamma,:gammaPoisson,:L2loglink,....)

Examples of use

data = HTBdata(y,x,param)
data = HTBdata(y,x,param,dates,fnames=names)
data = HTBdata(y,df[:,[:CAPE, :momentum ]],param,df.dates,fnames=df.names)
data = HTBdata(y,df[:,3:end],param)

Notes

When dates are provided, the data will be ordered chronologically (for cross-validation functions) unless the user has provided explicit training and validation sets.

HybridTreeBoosting.HTBindexes_from_dates — Function

HTBindexesfromdates(df::DataFrame,datesymbol::Symbol,firstdate::Date,nreestimate::Int)

Computes indexes of training set and test set for cumulative CV and pseudo-real-time forecasting exercises in time series and panel data. The function is inefficient for large datasets, but does not require the data to be sorted.

INPUTS

df DataFrame with dates and other variables

datesymbol symbol (or string) name of the date

first_date when the first training set ENDS (end date of the first training set)

n_reestimate every how many periods to re-estimate (update the training set)

OUTPUT indtraina,indtesta are arrays of arrays of indexes of train and test samples

Example of use

firstdate = Date("2017-12-31", Dates.DateFormat("y-m-d")) indtraina,indtesta = HTBindexesfromdates(df,:date,firstdate,12)

or as a named tuple with elements indtraina,indtesta)

t = HTBindexesfromdates(df,:date,first_date,12)

NOTES

Inefficient for large datasets

Fit and predict

HybridTreeBoosting.HTBfit — Function

HTBfit(data,param;cv_grid=[],cv_sparsity=:Auto,cv_depthppr=:Auto)

Fits HTBoost with with k-fold cross-validation of number of trees and depth, and possibly a few more models.

If param.modality is :fast or :fastest, fits one model, at param, and if needed a second where sharp splits are forced on features with high average values of τ. For param.modality=:accurate or :compromise, may then cross-validate the following hyperparamters:

Parameters for categorical features, if any.

depth in the range 1-6 (1-7 for :accurate) (starts with 3 and 5, then moves up or down based on results, breaking the loop as soon as the cv loss increases)

A penalization to encourage sparsity (fewer relevant features). Whether this is performed or not depends on the effective sample size (n and signal-to-noise ratio).

A model without projection pursuit nonlinear expansion of trees (only if modality=:accurate).

One or two models with column subsampling and slightly lower learning rate. One in :compromise, two in :accurate if the first reduced the cv loss.

A maximum of 10 models are fitted in modality = :accurate, and of 7 in :compromise.

The default range for depth can be replaced by providing a vector cv_grid, e.g.

HTBfit(data,param,cv_grid = [2,4,6])

The sparsity penalization can be de-activated by setting cv_sparsity=false, e.g.

    HTBfit(data,param,cv_sparsity=false)

If param.modality=:accurate, the learning rate lambda for all models is left at param.lambda (0.1 in default). If modality=:compromise, lambda=0.2 is used in cv, and the best model is then refitted with lambda = param.lambda.

Finally, all the estimated models considered are stacked, with weights chosen to minimize the cross-validated (original) loss. Unless modality=:accurate, this stacking will typically be equivalent to the best model.

Inputs

data::HTBdata

param::HTBparam

Optional inputs

cv_grid::Vector Defaul [2,3,5,6]. The code performs a search in the space depth in [2,3,5,6], trying to fit few models if possible. Provide a vector to over-ride (e.g. [2,4])

cv_sparsity Default :Auto. Set cvsparsity = true to guarantee search over sparsity penalization or cvsparsity=false to disactivate (and save computing time.) In :Auto, whether the cv is performed or not depends on :modality and on the n/p ratio. (Not implemented yet: it should ideally also depend on the signal-to-noise ratio).

cv_depthppr true to cv whether to remove projection pursuit regression. Default is true for modality in [:compromise,:accurate], else false.

cv_col_subsample Default:Auto. If true, fits one or two models with column subsampling.

Output (named tuple, or vector of named tuple for hurdle models)

indtest::Vector{Vector{Int}} indexes of validation samples

bestvalue::Float best value of depth in cv_grid

bestparam::SAMRTparam param for best model

ntrees::Int number of trees (best value of param.ntrees) for best model

loss::Float best cv loss

lossw::Float loss of stacked models

meanloss:Vector{Float} mean cv loss at bestvalue of param for param.ntrees = 1,2,....

stdeloss:Vector{Float} standard errror of cv loss at bestvalue of param for param.ntrees = 1,2,....

lossgrid::Vector{Float} cv loss for best tree size for each grid value

loglikdivide:Float loglikdivide. effective sample size = n/loglikvide. Roughly accounts for cross-correlation and serial correlation

HTBtrees::HTBoostTrees for the best cv value of param and ntrees

HTBtrees_a length(cv_grid) vector of HTBtrees

i vector (ntrees) of vectors (depth of each tree) of threshold features for best model

mu vector (ntrees) of vectors (depth of each tree) of threshold points for best model

tau vector (ntrees) of vectors (depth of each tree) of sigmoid parameters for best model

fi2 vector (ntrees) of vectors (depth of each tree) of feature importance, increase in R2 at each split, for best model

w length(cv_grid) vector of stacked weights

ratio_actual_max ratio of actual number of candidate features over potential maximum. Relevant if sparsevs=:On: indicates sparsevs should be switched off if too high (e.g. higher than 0.5).

problems true if there were computational problems in any of the models: NaN loss or loss jumping up

Notes

The following options for cross-validation are specified in param: randomizecv, nfold, sharevalidation, stderulestop

Examples of use:

param = HTBparam()
data   = HTBdata(y,x,param)
output = HTBfit(data,param)
ntrees = output.ntrees 
best_depth = output.bestvalue 

Example for hudle models (loss in [:hurdleGamma,:hurdleL2loglink,:hurdleL2])
ntrees_0    = output[1].ntrees   # number of trees for logistic regression, 0-not0
ntrees_not0 = output[2].ntrees   # number of trees for gamma or L2 loss

HybridTreeBoosting.HTBpredict — Function

HTBpredict(x,output)

Forecasts from HTBoost, for y or the natural parameter.

Inputs

x (n,p) DataFrame or Float matrix of forecast origins (type<:real) or p vector of forecast origin In the same format as the x given as input is HTBdata(y,x,...). May contain missing or NaN.

output output from HTBfit

Optional inputs

predict [:Ey], :Ey or :Egamma. :Ey returns the forecast of y, :Egamma returns the forecast of the natural parameter (e.g. logit(prob) for :logistic and E(log(y)|x) for :lognormal).

best_model [false] true to use only the single best model, false to use stacked weighted average

offset (n) vector, offset (or exposure), in terms of gamma (log exposure if the loss has a log-link)

cutoff_paralellel [20_000] if x has more than these rows, a parallellized algorithm is called (which is slower for few forecasts)

Output for standard models

yf (n) vector of forecasts of y (or, outside regression, of the natural parameter), or scalar forecast if n = 1

Output for hurdle models

yf (n) vector of forecasts of E(y|x) for the combined model

prob0 (n) vector of forecasts of prob(y=0|x)

yf_not0 (n) vector of forecasts of E(y|x,y /=0)

Output for loss = :multiclass

yf (n,num_class) matrix, yf[i,j] if the probability that observation i belongs to class j

class_values (numclas) vector, classvalue[j] is the value of y associated with class j

ymax (n) vector, ymax[i] is the class value with highest probability at observation i.

Examples of use

output = HTBfit(data,param)
yf     = HTBpredict(x_oos,output)
yf     = HTBpredict(x_oos,output,best_model=true)
yf     = HTBpredict(x_oos,output,offset = log.(exposure) )

yf,prob0,yf_not0 = HTBpredict(x_oos,output)  # for hurdle models. Or as a tuple t = HTBpredict(x_oos,output) 
yf,class_value,ymax = HTBpredict(x_oos,output)  # for multiclass. Or as a tuple t = HTBpredict(x_oos,output)

HybridTreeBoosting.HTBcv — Function

HTBcv(data,param,params_cv;internal_cv=true)

User's controlled cross-validation for HTBoost.

The recommended process for HTBoost is to use modality (:fastest, :fast, :compromise, :accurate) in HTBfit() rather than HTBcv(). The various modalities in HTBfit() internally control the most important hyperparameters, being as parsimonious as possible due to the high computational costs of HTB. HTBfit() also stacks the models, which is not done here. All modalities use early stopping to determine the number of trees, so ntrees should not be cv.

The function HTBcv() is provided for advanced users who want to fully control the cross-validation process and can incur the costs. HTBcv() has the option to set internal_cv=true, in which case cv is also performed internally (at increase cost).

A good use of HTBcv() is to cv a parameter that is not included in the modality (e.g. the strength of the smoothness prior), in conjunction with internal_cv = true. (At increase computational cost.)

Inputs

data HTBdata type

param HTBparam type. Includes number of folds, randomization (or block-cv) and, optionally, indexes of train and validation folds

params_cv Array of Dictionaries. Each dictionary contains the hyperparameters to be cv'ed (keys) and the values to be cv'ed (values).

Optional Inputs

internal_cv [true] If true, it will perform internal cv (as dictated by param.modality) for each element of the dictionary. This could be very slow. Set to false if you want to fully control the cv process. NOTE: if false, param.modality is irrelevant (set to :fast)

Output (named tuple)

bestindex Index of the best model in params_cv

bestparam param (type HTBparam) of the best model

output output of the best model

loss loss of the best model

output_a Array of outputs for all models

loss_a Array of losses for all models

Example of use: cv over varlntau (strength of smoothness prior), and also cv internally.

# Set up model and data 
param  = HTBparam(loss=:L2,modality=:accurate,nfold=4)  # number of folds, randomization (or block-cv) and, optionally, indexes of train and validation folds
data   = HTBdata(x,y,param)

# Specify hyperparameters and values for cv as a Dictionary. Here we cv only varlntau (strength of smoothness prior).
# The resuls is a one-dimensional array. More dimensions can be added in the same way. 
params_cv = [Dict(
    :varlntau => varlntau)    # strength of prior on log(tau) (smoothness). Default is 0.5^2. Smaller numbers are stronger priors.
    for
    varlntau in (0.25^2,0.5^1,1.0^2)
    ]

htbcv = HTBcv(data,param,params_cv)      # cv over dictionary and internally

# Some info about the best model 
bestindex = htbcv.bestindex   # params_cv[bestindex] is for best set of cv hyperparameters       
bestparam = htbcv.bestparam   

# predict using best model
yf    = HTBpredict(x_oos,htbcv.output)

Example of use: replace HTBfit() (not recommended) by setting internal_cv=false

# Set up model and data 
param  = HTBparam(loss=:L2,randomizecv=false,nfold=4)  # number of folds, randomization (or block-cv) and, optionally, indexes of train and validation folds
data   = HTBdata(x,y,param)

# Specify hyperparameters and values for cv as a Dictionary. Here we cv depth and varlntau (strength of smoothness prior).
# The resuls is a two-dimensional array. More dimensions can be added in the same way. 
params_cv = [Dict(
    :depth => depth,
    :varlntau => varlntau)    # strength of prior on log(tau) (smoothness). Default is 0.5^2. Smaller numbers are stronger priors.
    for
    depth in (2,4,6),
    varlntau in (0.25^2,0.5^1,1.0^2)
    ]

htbcv = HTBcv(data,param,params_cv,internal_cv=false)       

# Some info about the best model 
bestindex = htbcv.bestindex   # params_cv[bestindex] is for best set of cv hyperparameters       
bestparam = htbcv.bestparam   
output    = htbcv.output

# predict using best model
yf    = HTBpredict(x_oos,htbcv.output)

Post-estimation analysis

HybridTreeBoosting.HTBcoeff — Function

HTBcoeff(output;verbose=true)

Provides some information on constant coefficients for best model (in the form of a tuple.) For example, error variance for :L2, dispersion and dof for :t.

Inputs

output output from HTBfit

Output

coeff named tuple with information on fixed coefficients (e.g. variance for :L2, dispersion and dof for :t)

Example of use

output = HTBfit(data,param)
coeff  = HTBcoeff(output,verbose=false)

HybridTreeBoosting.HTBrelevance — Function

HTBrelevance(output,data::HTBdata;verbose=true,best_model=false)

Computes feature importance (summing to 100), defined by the relevance measure of Breiman et al. (1984), equation 10.42 in Hastie et al., "The Elements of Statistical Learning", second edition, except that the normalization is for sum = 100, not for largest = 100. Relevance is defined on the fit of the trees on pseudo-residuals. bestmodel=true for single model with lowest CV loss, bestmodel= false for weighted average (weights optimized by stacking)

Output

fnames::Vector{String} feature names, same order as in data

fi::Vector{Float} feature importance, same order as in data

fnames_sorted::Vector{String} feature names, sorted from highest to lowest importance

fi_sorted::Vector{Float} feature importance, sorted from highest to lowest

sortedindx::Vector{Int} feature indices, sorted from highest to lowest importance

Example of use

output = HTBfit(data,param)
fnames,fi,fnames_sorted,fi_sorted,sortedindx = HTBrelevance(output,data,verbose = false)
fnames,fi,fnames_sorted,fi_sorted,sortedindx = HTBrelevance(output,data,best_model=true)

or as a named tuple 

t = HTBrelevance(output,data,verbose = false)

HybridTreeBoosting.HTBpartialplot — Function

HTBpartialplot(data::HTBdata,output,features::Vector{Int64};,predict=:Egamma,best_model=false,other_xs::Vector=[],q1st=0.01,npoints=1000))

Partial dependence plot for selected features. Notice: Default is for natural parameter (gamma) rather than y. For feature i, computes gamma(xi) - gamma(xi=mean(xi)) for xi between q1st and 1-q1st quantile, with all other features at their mean.

Inputs

data::HTBdata

output

features::Vector{Int} position index (in data.x) of features to compute partial dependence plot for

other_xs::Vector{Float} (keyword), a size(data.x)[2] vector of values at which to evaluate the responses. [] Note: otherxs should be expressed in standardized units, i.e. for (xi-mean(xi))/std(xi)

q1st::Float (keyword) first quantile to compute, e.g. 0.001. Last quantile is 1-q1st. [0.01]

`npoints::Int' (keyword) number of points at which to evalute f(x). [1000]

Optional inputs

predict [:Egamma], :Ey or :Egamma. :Ey returns the impact on the forecast of y, :Egamma on the natural parameter.

best_model [false] true for single model with lowest CV loss, =false for weighted average (by stacking)

Output

q::Matrix (npoints,length(features)), values of xi at which f(xi) is evaluated

pdp::Matrix (npoints,length(features)), values of f(x_i)

Example of use

output = HTBfit(data,param)
q,pdp  = HTBpartialplot(data,output.HTBtrees,sortedindx[1,2],q1st=0.001)

or as a named tuple 

t      = HTBpartialplot(data,output.HTBtrees,sortedindx[1,2],q1st=0.001)

HybridTreeBoosting.HTBmarginaleffect — Function

HTBmarginaleffect(data::HTBdata,output,features::Vector{Int64};predict=:Egamma,best_model=false,other_xs::Vector =[],q1st=0.01,npoints=50,epsilon=0.02)

APPROXIMATE Computation of marginal effects using NUMERICAL derivatives (default ϵ=0.01)

Inputs

data::HTBdata

HTBtrees::HTBoostTrees

features::Vector{Int} position index (in data.x) of features to compute partial dependence plot for

other_xs::Vector{Float} (keyword), a size(data.x)[1] vector of values at which to evaluate the marginal effect. [] Note: otherxs should be expressed in standardized units, i.e. for (xi-mean(xi))/std(xi)

q1st::Float (keyword) first quantile to compute, e.g. 0.001. Last quantile is 1-q1st. [0.01]

`npoints::Int' (keyword) number of points at which to evalute df(xi)/dxi. [50]

`epsilon::Float' (keyword) epsilon for numerical derivative, [0.01]

Optional inputs

predict [:Egamma], :Ey or :Egamma. :Ey returns the impact on the forecast of y, :Egamma on the natural parameter.

best_model [false] true for single model with lowest CV loss, =false for weighted average (by stacking)

Output

q::Matrix (npoints,length(features)), values of xi at which f(xi) is evaluated, or vector if npoints = 1

d::Matrix (npoints,length(features)), values of marginal effects, or vector if npoints = 1

NOTE: Provisional! APPROXIMATE Computation of marginal effects using NUMERICAL derivatives. (Analytical derivatives are available)

NOTE: To compute marginal effect at one point x0 rather than over a grid, set npoints = 1 and other_xs = x0 (a p vector, p the number of features)

Example

output = HTBfit(data,param)
q,m    = HTBmarginaleffect(data,output.HTBtrees,[1,3])
t      = HTBmarginaleffect(data,output.HTBtrees,[1,3])   # as named tuple

Example

q,m  = HTBmarginaleffect(data,output.HTBtrees,[1,2,3,4],other_xs = zeros(p),npoints = 1)

Example

output = HTBfit(data,param)
fnames,fi,fnames_sorted,fi_sorted,sortedindx = HTBrelevance(output.HTBtrees,data,verbose=false)
q,m  = HTBmarginaleffect(data,output,sortedindx[1,2],q1st=0.001)

HybridTreeBoosting.HTBoutput — Function

HTBoutput(HTBtrees::HTBoostTrees;exclude_pp = true)

Output fitted parameters estimated from each tree, collected in matrices. Excluded projection pursuit regression parameters.

Inputs

The default exclude_pp does not give μ and τ for projection pursuit regression.

Output

i vector (ntrees) of vectors (depth of tree i) of threshold features

μ vector (ntrees) of vectors (depth of tree i) of threshold points

τ vector (ntrees) of vectors (depth of tree i) of smoothing parameters

fi2 vector (ntrees) of vectors (depth of tree i) of feature importance, increase in R2 at each split

m vector (ntrees) of vectors (depth of tree i) of threshold points for missing values

β vector (ntrees) of vectors (2^depth of tree i) of leaf coefficients (1st phase, excluding ppr)

Example of use

output = HTBfit(data,param)
i,μ,τ,fi2,m,β = HTBoutput(output.HTBtrees)
t = HTBoutput(output.HTBtrees)  # named tuple

HybridTreeBoosting.HTBweightedtau — Function

HTBweightedtau(output,data;verbose=true,best_model=false)

Computes weighted (by variance importance gain at each split) smoothing parameter τ for each feature, and for the entire model (features are averaged by variance importance) statistics for each feature, averaged over all trees. Sharp thresholds (τ=Inf) are bounded at 40. bestmodel=true for single model with lowest CV loss, bestmodel= false for weighted average (weights optimized by stacking)

Input

output output from HTBfit

data data input to HTBfit

Optional inputs

verbose [true] prints out the results to screen as DataFrame

max_tau [40] values of τ at or above this are set to Inf (hard splits) for plotting purposes

Output. Named tuple with the following fields

avgtau scalar, average importance weighted τ over all features (also weighted by variance importance)

gavgtau scalar, geometric average of importance weighted log τ over all features (also weighted by variance importance). NB: Default in printed output and plot. (Arguably the more informative measure.)

avgtau_a p-vector of avg importance weighted τ for each feature

df dataframe collecting avgtau_a information (only if verbose=true)

x_plot x-axis to plot sigmoid for gavgtau, in range [-2 2] for standardized feature

g_plot y-axis to plot sigmoid for gavgtau

fnames feature names

fi feature importance, summing to 100

fnames_sorted sorted feature names, from highest to lowest importance

fi_sorted sorted feature importance

sortedindx sorted feature indices

Example of use

output = HTBfit(data,param)
avgtau,gavgtau,avgtau_a,dftau,x_plot,g_plot = HTBweightedtau(output,data)
avgtau,gavgtau,avgtau_a,dftau,x_plot,g_plot = HTBweightedtau(output,data,verbose=false,plot_tau=false,best_model=true)

using Plots
plot(x_plot,g_plot,title="avg smoothness of splits",xlabel="standardized x",label=:none,legend=:bottomright)    

Or as a named tuple 

t = HTBweightedtau(output,data)
plot(t.x_plot,t.g_plot,title="avg smoothness of splits",xlabel="standardized x",label=:none,legend=:bottomright)

HybridTreeBoosting.HTBplot_tau — Function

HTBplot_tau(tau;mu=0,sigmoid=:sigmoidsqrt,range=2)

Produces x and y to plot sigmoid function for a given τ (typically the average τ for a feature).

Input

tau smoothing parameter τ

Optional inputs

mu [0] location parameter for sigmoid

sigmoid [sigmoidsqrt] :sigmoidsqrt or :sigmoidlogistic

range [2] x range for plot, [-range range]. x is standardized

max_tau [40] tau at which a hard split is assumed

Output

x_plot x-axis to plot sigmoid for tau, in range [-2 2] for standardized feature

g_plot y-axis to plot sigmoid for tau

Note

tau ≥ 40 is interpreted as a hard split.

Example of use

output = HTBfit(data,param)
avgtau,gavgtau,avgtau_a,dftau,x_plot,g_plot = HTBweightedtau(output,data)
x_plot,g_plot = HTBplot_tau(avgtau[1])     # tau of first feature
using Plots
plot(x_plot,g_plot,title="avg tau of feature 1",xlabel="standardized x",label=:none,legend=:bottomright)

or as a named tuple

t = HTBplot_tau(avgtau[1])     # tau of first feature
plot(t.x_plot,t.g_plot,title="avg tau of feature 1",xlabel="standardized x",label=:none,legend=:bottomright)

HybridTreeBoosting.HTBplot_ppr — Function

HTBplot_ppr(output;which_tree=1)

Visualize impact of projection pursuit for an individual tree.

Example of use

yf1,yf0,tau = HTBplot_ppr(output,which_tree=1)
plot(yf0,yf1)

where yf0 is the standardized prediction from the tree, and yf1 is the (non-standardized) prediction after ppr